+++ tilebox-datasets/tilebox/datasets/datasets/v1/datasets_pb2.py
@@ -28,7 +28,7 @@ from tilebox.datasets.datasets.v1 import dataset_type_pb2 as datasets_dot_v1_dot
 from tilebox.datasets.tilebox.v1 import id_pb2 as tilebox_dot_v1_dot_id__pb2
 
 
-DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\x1a\x64\x61tasets/v1/datasets.proto\x12\x0b\x64\x61tasets.v1\x1a\x1b\x62uf/validate/validate.proto\x1a\x16\x64\x61tasets/v1/core.proto\x1a\x1e\x64\x61tasets/v1/dataset_type.proto\x1a\x13tilebox/v1/id.proto\"\xa9\x01\n\x14\x43reateDatasetRequest\x12\x1b\n\x04name\x18\x01 \x01(\tB\x07\xbaH\x04r\x02\x10\x01R\x04name\x12\x34\n\x04type\x18\x02 \x01(\x0b\x32\x18.datasets.v1.DatasetTypeB\x06\xbaH\x03\xc8\x01\x01R\x04type\x12\x18\n\x07summary\x18\x03 \x01(\tR\x07summary\x12$\n\tcode_name\x18\x04 \x01(\tB\x07\xbaH\x04r\x02\x10\x01R\x08\x63odeName\"\x86\x01\n\x11GetDatasetRequest\x12>\n\x04slug\x18\x01 \x01(\tB*\xbaH\'r%2#^[a-zA-Z0-9_]+(?:\\.[a-zA-Z0-9_]+)*$R\x04slug\x12\x1e\n\x02id\x18\x02 \x01(\x0b\x32\x0e.tilebox.v1.IDR\x02id:\x11\xbaH\x0e\"\x0c\n\x04slug\n\x02id\x10\x01\"\xb2\x01\n\x14UpdateDatasetRequest\x12&\n\x02id\x18\x01 \x01(\x0b\x32\x0e.tilebox.v1.IDB\x06\xbaH\x03\xc8\x01\x01R\x02id\x12\x1b\n\x04name\x18\x02 \x01(\tB\x07\xbaH\x04r\x02\x10\x01R\x04name\x12\x34\n\x04type\x18\x03 \x01(\x0b\x32\x18.datasets.v1.DatasetTypeB\x06\xbaH\x03\xc8\x01\x01R\x04type\x12\x1f\n\x07summary\x18\x04 \x01(\tB\x05\xaa\x01\x02\x08\x01R\x07summary\"t\n\nClientInfo\x12\x12\n\x04name\x18\x01 \x01(\tR\x04name\x12 \n\x0b\x65nvironment\x18\x02 \x01(\tR\x0b\x65nvironment\x12\x30\n\x08packages\x18\x03 \x03(\x0b\x32\x14.datasets.v1.PackageR\x08packages\"7\n\x07Package\x12\x12\n\x04name\x18\x01 \x01(\tR\x04name\x12\x18\n\x07version\x18\x02 \x01(\tR\x07version\"k\n\x1fUpdateDatasetDescriptionRequest\x12&\n\x02id\x18\x01 \x01(\x0b\x32\x0e.tilebox.v1.IDB\x06\xbaH\x03\xc8\x01\x01R\x02id\x12 \n\x0b\x64\x65scription\x18\x02 \x01(\tR\x0b\x64\x65scription\">\n\x14\x44\x65leteDatasetRequest\x12&\n\x02id\x18\x01 \x01(\x0b\x32\x0e.tilebox.v1.IDB\x06\xbaH\x03\xc8\x01\x01R\x02id\"1\n\x15\x44\x65leteDatasetResponse\x12\x18\n\x07trashed\x18\x01 \x01(\x08R\x07trashed\"O\n\x13ListDatasetsRequest\x12\x38\n\x0b\x63lient_info\x18\x01 \x01(\x0b\x32\x17.datasets.v1.ClientInfoR\nclientInfo\"\x86\x02\n\x14ListDatasetsResponse\x12\x30\n\x08\x64\x61tasets\x18\x01 \x03(\x0b\x32\x14.datasets.v1.DatasetR\x08\x64\x61tasets\x12\x31\n\x06groups\x18\x02 \x03(\x0b\x32\x19.datasets.v1.DatasetGroupR\x06groups\x12%\n\x0eserver_message\x18\x03 \x01(\tR\rserverMessage\x12%\n\x0eowned_datasets\x18\x04 \x01(\x03R\rownedDatasets\x12;\n\x16maximum_owned_datasets\x18\x05 \x01(\x03\x42\x05\xaa\x01\x02\x08\x01R\x14maximumOwnedDatasets2\x81\x04\n\x0e\x44\x61tasetService\x12J\n\rCreateDataset\x12!.datasets.v1.CreateDatasetRequest\x1a\x14.datasets.v1.Dataset\"\x00\x12\x44\n\nGetDataset\x12\x1e.datasets.v1.GetDatasetRequest\x1a\x14.datasets.v1.Dataset\"\x00\x12J\n\rUpdateDataset\x12!.datasets.v1.UpdateDatasetRequest\x1a\x14.datasets.v1.Dataset\"\x00\x12`\n\x18UpdateDatasetDescription\x12,.datasets.v1.UpdateDatasetDescriptionRequest\x1a\x14.datasets.v1.Dataset\"\x00\x12X\n\rDeleteDataset\x12!.datasets.v1.DeleteDatasetRequest\x1a\".datasets.v1.DeleteDatasetResponse\"\x00\x12U\n\x0cListDatasets\x12 .datasets.v1.ListDatasetsRequest\x1a!.datasets.v1.ListDatasetsResponse\"\x00\x42r\n\x0f\x63om.datasets.v1B\rDatasetsProtoP\x01\xa2\x02\x03\x44XX\xaa\x02\x0b\x44\x61tasets.V1\xca\x02\x0b\x44\x61tasets\\V1\xe2\x02\x17\x44\x61tasets\\V1\\GPBMetadata\xea\x02\x0c\x44\x61tasets::V1\x92\x03\x02\x08\x02\x62\x08\x65\x64itionsp\xe8\x07')
+DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\x1a\x64\x61tasets/v1/datasets.proto\x12\x0b\x64\x61tasets.v1\x1a\x1b\x62uf/validate/validate.proto\x1a\x16\x64\x61tasets/v1/core.proto\x1a\x1e\x64\x61tasets/v1/dataset_type.proto\x1a\x13tilebox/v1/id.proto\"\xfe\x01\n\x14\x43reateDatasetRequest\x12 \n\x04name\x18\x01 \x01(\tB\x0c\xbaH\tr\x04\x10\x01\x18\x64\xc8\x01\x01R\x04name\x12\x34\n\x04type\x18\x02 \x01(\x0b\x32\x18.datasets.v1.DatasetTypeB\x06\xbaH\x03\xc8\x01\x01R\x04type\x12#\n\x07summary\x18\x03 \x01(\tB\t\xbaH\x06r\x04\x18\xd0\x86\x03R\x07summary\x12<\n\tcode_name\x18\x04 \x01(\tB\x1f\xbaH\x1cr\x17\x10\x01\x18\x64\x32\x11^[a-z][a-z0-9_]*$\xc8\x01\x01R\x08\x63odeName\x12+\n\x0b\x64\x65scription\x18\x05 \x01(\tB\t\xbaH\x06r\x04\x18\xc0\x84=R\x0b\x64\x65scription\"\x86\x01\n\x11GetDatasetRequest\x12>\n\x04slug\x18\x01 \x01(\tB*\xbaH\'r%2#^[a-zA-Z0-9_]+(?:\\.[a-zA-Z0-9_]+)*$R\x04slug\x12\x1e\n\x02id\x18\x02 \x01(\x0b\x32\x0e.tilebox.v1.IDR\x02id:\x11\xbaH\x0e\"\x0c\n\x04slug\n\x02id\x10\x01\"\xb3\x02\n\x14UpdateDatasetRequest\x12&\n\x02id\x18\x01 \x01(\x0b\x32\x0e.tilebox.v1.IDB\x06\xbaH\x03\xc8\x01\x01R\x02id\x12\"\n\x04name\x18\x02 \x01(\tB\x0e\xaa\x01\x02\x08\x01\xbaH\x06r\x04\x10\x01\x18\x64R\x04name\x12\x33\n\x04type\x18\x03 \x01(\x0b\x32\x18.datasets.v1.DatasetTypeB\x05\xaa\x01\x02\x08\x01R\x04type\x12(\n\x07summary\x18\x04 \x01(\tB\x0e\xaa\x01\x02\x08\x01\xbaH\x06r\x04\x18\xd0\x86\x03R\x07summary\x12>\n\tcode_name\x18\x05 \x01(\tB!\xaa\x01\x02\x08\x01\xbaH\x19r\x17\x10\x01\x18\x64\x32\x11^[a-z][a-z0-9_]*$R\x08\x63odeName\x12\x30\n\x0b\x64\x65scription\x18\x06 \x01(\tB\x0e\xaa\x01\x02\x08\x01\xbaH\x06r\x04\x18\xc0\x84=R\x0b\x64\x65scription\"t\n\nClientInfo\x12\x12\n\x04name\x18\x01 \x01(\tR\x04name\x12 \n\x0b\x65nvironment\x18\x02 \x01(\tR\x0b\x65nvironment\x12\x30\n\x08packages\x18\x03 \x03(\x0b\x32\x14.datasets.v1.PackageR\x08packages\"7\n\x07Package\x12\x12\n\x04name\x18\x01 \x01(\tR\x04name\x12\x18\n\x07version\x18\x02 \x01(\tR\x07version\">\n\x14\x44\x65leteDatasetRequest\x12&\n\x02id\x18\x01 \x01(\x0b\x32\x0e.tilebox.v1.IDB\x06\xbaH\x03\xc8\x01\x01R\x02id\"1\n\x15\x44\x65leteDatasetResponse\x12\x18\n\x07trashed\x18\x01 \x01(\x08R\x07trashed\"O\n\x13ListDatasetsRequest\x12\x38\n\x0b\x63lient_info\x18\x01 \x01(\x0b\x32\x17.datasets.v1.ClientInfoR\nclientInfo\"\x86\x02\n\x14ListDatasetsResponse\x12\x30\n\x08\x64\x61tasets\x18\x01 \x03(\x0b\x32\x14.datasets.v1.DatasetR\x08\x64\x61tasets\x12\x31\n\x06groups\x18\x02 \x03(\x0b\x32\x19.datasets.v1.DatasetGroupR\x06groups\x12%\n\x0eserver_message\x18\x03 \x01(\tR\rserverMessage\x12%\n\x0eowned_datasets\x18\x04 \x01(\x03R\rownedDatasets\x12;\n\x16maximum_owned_datasets\x18\x05 \x01(\x03\x42\x05\xaa\x01\x02\x08\x01R\x14maximumOwnedDatasets2\x9f\x03\n\x0e\x44\x61tasetService\x12J\n\rCreateDataset\x12!.datasets.v1.CreateDatasetRequest\x1a\x14.datasets.v1.Dataset\"\x00\x12\x44\n\nGetDataset\x12\x1e.datasets.v1.GetDatasetRequest\x1a\x14.datasets.v1.Dataset\"\x00\x12J\n\rUpdateDataset\x12!.datasets.v1.UpdateDatasetRequest\x1a\x14.datasets.v1.Dataset\"\x00\x12X\n\rDeleteDataset\x12!.datasets.v1.DeleteDatasetRequest\x1a\".datasets.v1.DeleteDatasetResponse\"\x00\x12U\n\x0cListDatasets\x12 .datasets.v1.ListDatasetsRequest\x1a!.datasets.v1.ListDatasetsResponse\"\x00\x42r\n\x0f\x63om.datasets.v1B\rDatasetsProtoP\x01\xa2\x02\x03\x44XX\xaa\x02\x0b\x44\x61tasets.V1\xca\x02\x0b\x44\x61tasets\\V1\xe2\x02\x17\x44\x61tasets\\V1\\GPBMetadata\xea\x02\x0c\x44\x61tasets::V1\x92\x03\x02\x08\x02\x62\x08\x65\x64itionsp\xe8\x07')
 
 _globals = globals()
 _builder.BuildMessageAndEnumDescriptors(DESCRIPTOR, _globals)
@@ -37,11 +37,15 @@ if not _descriptor._USE_C_DESCRIPTORS:
   _globals['DESCRIPTOR']._loaded_options = None
   _globals['DESCRIPTOR']._serialized_options = b'\n\017com.datasets.v1B\rDatasetsProtoP\001\242\002\003DXX\252\002\013Datasets.V1\312\002\013Datasets\\V1\342\002\027Datasets\\V1\\GPBMetadata\352\002\014Datasets::V1\222\003\002\010\002'
   _globals['_CREATEDATASETREQUEST'].fields_by_name['name']._loaded_options = None
-  _globals['_CREATEDATASETREQUEST'].fields_by_name['name']._serialized_options = b'\272H\004r\002\020\001'
+  _globals['_CREATEDATASETREQUEST'].fields_by_name['name']._serialized_options = b'\272H\tr\004\020\001\030d\310\001\001'
   _globals['_CREATEDATASETREQUEST'].fields_by_name['type']._loaded_options = None
   _globals['_CREATEDATASETREQUEST'].fields_by_name['type']._serialized_options = b'\272H\003\310\001\001'
+  _globals['_CREATEDATASETREQUEST'].fields_by_name['summary']._loaded_options = None
+  _globals['_CREATEDATASETREQUEST'].fields_by_name['summary']._serialized_options = b'\272H\006r\004\030\320\206\003'
   _globals['_CREATEDATASETREQUEST'].fields_by_name['code_name']._loaded_options = None
-  _globals['_CREATEDATASETREQUEST'].fields_by_name['code_name']._serialized_options = b'\272H\004r\002\020\001'
+  _globals['_CREATEDATASETREQUEST'].fields_by_name['code_name']._serialized_options = b'\272H\034r\027\020\001\030d2\021^[a-z][a-z0-9_]*$\310\001\001'
+  _globals['_CREATEDATASETREQUEST'].fields_by_name['description']._loaded_options = None
+  _globals['_CREATEDATASETREQUEST'].fields_by_name['description']._serialized_options = b'\272H\006r\004\030\300\204='
   _globals['_GETDATASETREQUEST'].fields_by_name['slug']._loaded_options = None
   _globals['_GETDATASETREQUEST'].fields_by_name['slug']._serialized_options = b'\272H\'r%2#^[a-zA-Z0-9_]+(?:\\.[a-zA-Z0-9_]+)*$'
   _globals['_GETDATASETREQUEST']._loaded_options = None
@@ -49,37 +53,37 @@ if not _descriptor._USE_C_DESCRIPTORS:
   _globals['_UPDATEDATASETREQUEST'].fields_by_name['id']._loaded_options = None
   _globals['_UPDATEDATASETREQUEST'].fields_by_name['id']._serialized_options = b'\272H\003\310\001\001'
   _globals['_UPDATEDATASETREQUEST'].fields_by_name['name']._loaded_options = None
-  _globals['_UPDATEDATASETREQUEST'].fields_by_name['name']._serialized_options = b'\272H\004r\002\020\001'
+  _globals['_UPDATEDATASETREQUEST'].fields_by_name['name']._serialized_options = b'\252\001\002\010\001\272H\006r\004\020\001\030d'
   _globals['_UPDATEDATASETREQUEST'].fields_by_name['type']._loaded_options = None
-  _globals['_UPDATEDATASETREQUEST'].fields_by_name['type']._serialized_options = b'\272H\003\310\001\001'
+  _globals['_UPDATEDATASETREQUEST'].fields_by_name['type']._serialized_options = b'\252\001\002\010\001'
   _globals['_UPDATEDATASETREQUEST'].fields_by_name['summary']._loaded_options = None
-  _globals['_UPDATEDATASETREQUEST'].fields_by_name['summary']._serialized_options = b'\252\001\002\010\001'
-  _globals['_UPDATEDATASETDESCRIPTIONREQUEST'].fields_by_name['id']._loaded_options = None
-  _globals['_UPDATEDATASETDESCRIPTIONREQUEST'].fields_by_name['id']._serialized_options = b'\272H\003\310\001\001'
+  _globals['_UPDATEDATASETREQUEST'].fields_by_name['summary']._serialized_options = b'\252\001\002\010\001\272H\006r\004\030\320\206\003'
+  _globals['_UPDATEDATASETREQUEST'].fields_by_name['code_name']._loaded_options = None
+  _globals['_UPDATEDATASETREQUEST'].fields_by_name['code_name']._serialized_options = b'\252\001\002\010\001\272H\031r\027\020\001\030d2\021^[a-z][a-z0-9_]*$'
+  _globals['_UPDATEDATASETREQUEST'].fields_by_name['description']._loaded_options = None
+  _globals['_UPDATEDATASETREQUEST'].fields_by_name['description']._serialized_options = b'\252\001\002\010\001\272H\006r\004\030\300\204='
   _globals['_DELETEDATASETREQUEST'].fields_by_name['id']._loaded_options = None
   _globals['_DELETEDATASETREQUEST'].fields_by_name['id']._serialized_options = b'\272H\003\310\001\001'
   _globals['_LISTDATASETSRESPONSE'].fields_by_name['maximum_owned_datasets']._loaded_options = None
   _globals['_LISTDATASETSRESPONSE'].fields_by_name['maximum_owned_datasets']._serialized_options = b'\252\001\002\010\001'
   _globals['_CREATEDATASETREQUEST']._serialized_start=150
-  _globals['_CREATEDATASETREQUEST']._serialized_end=319
-  _globals['_GETDATASETREQUEST']._serialized_start=322
-  _globals['_GETDATASETREQUEST']._serialized_end=456
-  _globals['_UPDATEDATASETREQUEST']._serialized_start=459
-  _globals['_UPDATEDATASETREQUEST']._serialized_end=637
-  _globals['_CLIENTINFO']._serialized_start=639
-  _globals['_CLIENTINFO']._serialized_end=755
-  _globals['_PACKAGE']._serialized_start=757
-  _globals['_PACKAGE']._serialized_end=812
-  _globals['_UPDATEDATASETDESCRIPTIONREQUEST']._serialized_start=814
-  _globals['_UPDATEDATASETDESCRIPTIONREQUEST']._serialized_end=921
-  _globals['_DELETEDATASETREQUEST']._serialized_start=923
-  _globals['_DELETEDATASETREQUEST']._serialized_end=985
-  _globals['_DELETEDATASETRESPONSE']._serialized_start=987
-  _globals['_DELETEDATASETRESPONSE']._serialized_end=1036
-  _globals['_LISTDATASETSREQUEST']._serialized_start=1038
-  _globals['_LISTDATASETSREQUEST']._serialized_end=1117
-  _globals['_LISTDATASETSRESPONSE']._serialized_start=1120
-  _globals['_LISTDATASETSRESPONSE']._serialized_end=1382
-  _globals['_DATASETSERVICE']._serialized_start=1385
-  _globals['_DATASETSERVICE']._serialized_end=1898
+  _globals['_CREATEDATASETREQUEST']._serialized_end=404
+  _globals['_GETDATASETREQUEST']._serialized_start=407
+  _globals['_GETDATASETREQUEST']._serialized_end=541
+  _globals['_UPDATEDATASETREQUEST']._serialized_start=544
+  _globals['_UPDATEDATASETREQUEST']._serialized_end=851
+  _globals['_CLIENTINFO']._serialized_start=853
+  _globals['_CLIENTINFO']._serialized_end=969
+  _globals['_PACKAGE']._serialized_start=971
+  _globals['_PACKAGE']._serialized_end=1026
+  _globals['_DELETEDATASETREQUEST']._serialized_start=1028
+  _globals['_DELETEDATASETREQUEST']._serialized_end=1090
+  _globals['_DELETEDATASETRESPONSE']._serialized_start=1092
+  _globals['_DELETEDATASETRESPONSE']._serialized_end=1141
+  _globals['_LISTDATASETSREQUEST']._serialized_start=1143
+  _globals['_LISTDATASETSREQUEST']._serialized_end=1222
+  _globals['_LISTDATASETSRESPONSE']._serialized_start=1225
+  _globals['_LISTDATASETSRESPONSE']._serialized_end=1487
+  _globals['_DATASETSERVICE']._serialized_start=1490
+  _globals['_DATASETSERVICE']._serialized_end=1905
 # @@protoc_insertion_point(module_scope)